requirements.txt
python-dotenv
pydantic
langgraph
langchain
openai
gspread
oauth2client
google-api-python-client